# lease: do lease pile-up reduction in the background