@mgates 2018-05-04
I'll have some more realistic test examples on Monday.