@cosgroveb 2018-05-07
This addresses #9496