@mgates 2018-05-25
Ok - so I'm not sure what the best way to demonstarate this in a real-worldish way is, but here's what I did:

1) set up the test cluster with `goreman start`
2) put 1.9 million leases into etcd
3) Ran this snippet which tries to make a new lease every half second, and kills the leader after 10 seconds
```
while true; do ETCDCTL_API=3 etcdctl --command-timeout 0.2s --endpoints "http://127.0.0.1:2379,http://127.0.0.1:22379,http://127.0.0.1:32379" lease grant  60  > /dev/null ; echo `date` $? ; sleep 0.5; done & sleep 10; pkill -f  -- "[-]-listen-peer-urls $(ETCDCTL_API=2 etcdctl  member list  | grep isLeader=true | cut -f 3 -d "="| cut -f 1 -d " ")"
```

On the branch it had 1 timeout exceed error, but on master it 60, over the 61 seconds it took before it recovered.

Sorry it took so long - if these changes looks plausible to you, we can try to test it against our live, large workloads.