@mgates 2018-06-04
Hey folks - just wanted to check in about this - does it seem ok, at least in theory? If so, happy to work more on it.