@xiang90 2018-06-04
@mgates 

Can you:

1. describe what problem (in which case what can happen and how this PR can help ) this PR solves
2. move the sorting change into another PR

I can understand this PR as it is, but I guess it is easier for other people to review if you do what I suggested.