@xiang90 2018-06-06
@mgates What happens when a lease is revoked while the go routine created in the promote is still running? is there a risk that the copy of the lease will be re-add into the heap?