@mgates 2018-06-07
That's a good question - I'm pretty sure that it would get updated, but since the revoke method deletes it from the lease map and deletes the appropriate keys, that wouldn't actually be a problem, and the lease would get cleaned up eventually.That's a good question - I'm pretty sure that it would get updated, but since the revoke method deletes it from the lease map and deletes the appropriate keys, that wouldn't actually be a problem, and the lease would get cleaned up eventually.

We'll write a test to confirm, though