@mgates 2018-06-07
Hey there, we looked into this more, and it doesn't seem possible to write a automated test of this behavior, we did some manual work and added some sleeps and feel good about the behavior - the lease TTL will get refreshed and possibly rewritten for the pile up avoidance, it'll get added to the heap, and will be ignored once the expiry checker gets to it, because it isn't in the lease map.